Getting rid of vecs in input and trying to make commands a class
